Constraints

Work-in-progress

This page of the docs is still a work-in-progress. Check back later!

GCPDecompositions.GCPConstraints — Module

Constraints for Generalized CP Decomposition.

GCPDecompositions.GCPConstraints.AbstractConstraint — Type

AbstractConstraint

Abstract type for GCP constraints on the factor matrices U = (U[1],...,U[N]).

GCPDecompositions.GCPConstraints.LowerBound — Type

LowerBound(value::Real)

Lower-bound constraint on the entries of the factor matrices U = (U[1],...,U[N]), i.e., U[i][j,k] >= value.