GCPDecompositions: Generalized CP Decompositions

Documentation for GCPDecompositions.

👋 This package provides research code and work is ongoing. If you are interested in using it in your own research, I'd love to hear from you and collaborate! Feel free to write: hong@udel.edu

Please cite the following papers for this technique:

David Hong, Tamara G. Kolda, Jed A. Duersch. "Generalized Canonical Polyadic Tensor Decomposition", SIAM Review 62:133-163, 2020. https://doi.org/10.1137/18M1203626 https://arxiv.org/abs/1808.07452

Tamara G. Kolda, David Hong. "Stochastic Gradients for Large-Scale Tensor Decomposition", SIAM Journal on Mathematics of Data Science 2:1066-1095, 2020. https://doi.org/10.1137/19M1266265 https://arxiv.org/abs/1906.01687

In BibTeX form:

@Article{hkd2020gcp,
  title =        "Generalized Canonical Polyadic Tensor Decomposition",
  author =       "David Hong and Tamara G. Kolda and Jed A. Duersch",
  journal =      "{SIAM} Review",
  year =         "2020",
  volume =       "62",
  number =       "1",
  pages =        "133--163",
  DOI =          "10.1137/18M1203626",
}

@Article{kh2020sgf,
  title =        "Stochastic Gradients for Large-Scale Tensor Decomposition",
  author =       "Tamara G. Kolda and David Hong",
  journal =      "{SIAM} Journal on Mathematics of Data Science",
  year =         "2020",
  volume =       "2",
  number =       "4",
  pages =        "1066--1095",
  DOI =          "10.1137/19M1266265",
}

Docstrings

GCPDecompositions.GCPDecompositions

GCPDecompositions.AbstractLoss

GCPDecompositions.CPD

GCPDecompositions.LeastSquaresLoss

GCPDecompositions.NonnegativeLeastSquaresLoss

GCPDecompositions.PoissonLogLoss

GCPDecompositions.PoissonLoss

GCPDecompositions.UserDefinedLoss

GCPDecompositions.gcp

GCPDecompositions.GCPDecompositions — Module

Generalized CP Decomposition module. Provides approximate CP tensor decomposition with respect to general losses.

GCPDecompositions.AbstractLoss — Type

AbstractLoss

Abstract type for GCP loss functions $f(x,m)$, where $x$ is the data entry and $m$ is the model entry.

Concrete types ConcreteLoss <: AbstractLoss should implement:

value(loss::ConcreteLoss, x, m) that computes the value of the loss function $f(x,m)$

deriv(loss::ConcreteLoss, x, m) that computes the value of the partial derivative $\partial_m f(x,m)$ with respect to $m$

domain(loss::ConcreteLoss) that returns an Interval from IntervalSets.jl defining the domain for $m$

GCPDecompositions.CPD — Type

CPD

Tensor decomposition type for the canonical polyadic decompositions (CPD) of a tensor (i.e., a multi-dimensional array) A. This is the return type of gcp(_), the corresponding tensor decomposition function.

If M::CPD is the decomposition object, the weights λ and the factor matrices U = (U[1],...,U[N]) can be obtained via M.λ and M.U, such that A = Σ_j λ[j] U[1][:,j] ∘ ⋯ ∘ U[N][:,j].

GCPDecompositions.LeastSquaresLoss — Type

LeastSquaresLoss()

Loss corresponding to conventional CP decomposition. Corresponds to a statistical assumption of Gaussian data X with mean given by the low-rank model tensor M.

Distribution: $x_i \sim \mathcal{N}(\mu_i, \sigma)$

Link function: $m_i = \mu_i$

Loss function: $f(x,m) = (x-m)^2$

Domain: $m \in \mathbb{R}$

GCPDecompositions.NonnegativeLeastSquaresLoss — Type

NonnegativeLeastSquaresLoss()

Loss corresponding to nonnegative CP decomposition. Corresponds to a statistical assumption of Gaussian data X with nonnegative mean given by the low-rank model tensor M.

Distribution: $x_i \sim \mathcal{N}(\mu_i, \sigma)$

Link function: $m_i = \mu_i$

Loss function: $f(x,m) = (x-m)^2$

Domain: $m \in [0, \infty)$

GCPDecompositions.PoissonLogLoss — Type

PoissonLogLoss()

Loss corresponding to a statistical assumption of Poisson data X with log-rate given by the low-rank model tensor M.

Distribution: $x_i \sim \operatorname{Poisson}(\lambda_i)$

Link function: $m_i = \log \lambda_i$

Loss function: $f(x,m) = e^m - x m$

Domain: $m \in \mathbb{R}$

GCPDecompositions.PoissonLoss — Type

PoissonLoss(eps::Real = 1e-10)

Loss corresponding to a statistical assumption of Poisson data X with rate given by the low-rank model tensor M.

Distribution: $x_i \sim \operatorname{Poisson}(\lambda_i)$

Link function: $m_i = \lambda_i$

Loss function: $f(x,m) = m - x \log(m + \epsilon)$

Domain: $m \in [0, \infty)$

GCPDecompositions.UserDefinedLoss — Type

UserDefinedLoss

Type for user-defined loss functions $f(x,m)$, where $x$ is the data entry and $m$ is the model entry.

Contains three fields:

func::Function : function that evaluates the loss function $f(x,m)$

deriv::Function : function that evaluates the partial derivative $\partial_m f(x,m)$ with respect to $m$

domain::Interval : Interval from IntervalSets.jl defining the domain for $m$

The constructor is UserDefinedLoss(func; deriv, domain). If not provided,

deriv is automatically computed from func using forward-mode automatic differentiation

domain gets a default value of Interval(-Inf, +Inf)

GCPDecompositions.gcp — Function

gcp(X::Array, r, loss = LeastSquaresLoss()) -> CPD

Compute an approximate rank-r CP decomposition of the tensor X with respect to the loss function loss and return a CPD object. Conventional CP corresponds to the default LeastSquaresLoss().

If the LossFunctions.jl package is also loaded, loss can also be a loss function from that package. Check GCPDecompositions.LossFunctionsExt.SupportedLosses to see what losses are supported.

See also: CPD, AbstractLoss.