GCPDecompositions: Generalized CP Decompositions

Documentation for GCPDecompositions.

👋 This package provides research code and work is ongoing. If you are interested in using it in your own research, I'd love to hear from you and collaborate! Feel free to write: hong@udel.edu

Please cite the following papers for this technique:

David Hong, Tamara G. Kolda, Jed A. Duersch. "Generalized Canonical Polyadic Tensor Decomposition", SIAM Review 62:133-163, 2020. https://doi.org/10.1137/18M1203626 https://arxiv.org/abs/1808.07452

Tamara G. Kolda, David Hong. "Stochastic Gradients for Large-Scale Tensor Decomposition", SIAM Journal on Mathematics of Data Science 2:1066-1095, 2020. https://doi.org/10.1137/19M1266265 https://arxiv.org/abs/1906.01687

In BibTeX form:

@Article{hkd2020gcp,
  title =        "Generalized Canonical Polyadic Tensor Decomposition",
  author =       "David Hong and Tamara G. Kolda and Jed A. Duersch",
  journal =      "{SIAM} Review",
  year =         "2020",
  volume =       "62",
  number =       "1",
  pages =        "133--163",
  DOI =          "10.1137/18M1203626",
}

@Article{kh2020sgf,
  title =        "Stochastic Gradients for Large-Scale Tensor Decomposition",
  author =       "Tamara G. Kolda and David Hong",
  journal =      "{SIAM} Journal on Mathematics of Data Science",
  year =         "2020",
  volume =       "2",
  number =       "4",
  pages =        "1066--1095",
  DOI =          "10.1137/19M1266265",
}

Docstrings

GCPDecompositions.GCPDecompositions

GCPDecompositions.AbstractLoss

GCPDecompositions.BernoulliLogitLoss

GCPDecompositions.BernoulliOddsLoss

GCPDecompositions.BetaDivergenceLoss

GCPDecompositions.CPD

GCPDecompositions.GammaLoss

GCPDecompositions.HuberLoss

GCPDecompositions.LeastSquaresLoss

GCPDecompositions.NegativeBinomialOddsLoss

GCPDecompositions.NonnegativeLeastSquaresLoss

GCPDecompositions.PoissonLogLoss

GCPDecompositions.PoissonLoss

GCPDecompositions.RayleighLoss

GCPDecompositions.UserDefinedLoss

GCPDecompositions._checked_khatrirao_dims

GCPDecompositions._checked_mttkrp_dims

GCPDecompositions.create_mttkrp_buffer

GCPDecompositions.gcp

GCPDecompositions.khatrirao

GCPDecompositions.khatrirao!

GCPDecompositions.mttkrp

GCPDecompositions.mttkrp!

GCPDecompositions.GCPDecompositions — Module

Generalized CP Decomposition module. Provides approximate CP tensor decomposition with respect to general losses.

GCPDecompositions.AbstractLoss — Type

AbstractLoss

Abstract type for GCP loss functions $f(x,m)$, where $x$ is the data entry and $m$ is the model entry.

Concrete types ConcreteLoss <: AbstractLoss should implement:

value(loss::ConcreteLoss, x, m) that computes the value of the loss function $f(x,m)$

deriv(loss::ConcreteLoss, x, m) that computes the value of the partial derivative $\partial_m f(x,m)$ with respect to $m$

domain(loss::ConcreteLoss) that returns an Interval from IntervalSets.jl defining the domain for $m$

GCPDecompositions.BernoulliLogitLoss — Type

BernoulliLogitLoss(eps::Real = 1e-10)

Loss corresponding to the statistical assumption of Bernouli data X with log odds-success rate given by the low-rank model tensor M

Distribution: $x_i \sim \operatorname{Bernouli}(\rho_i)$

Link function: $m_i = \log(\frac{\rho_i}{1 - \rho_i})$

Loss function: $f(x, m) = \log(1 + e^m) - xm$

Domain: $m \in \mathbb{R}$

GCPDecompositions.BernoulliOddsLoss — Type

BernoulliOddsLoss(eps::Real = 1e-10)

Loss corresponding to the statistical assumption of Bernouli data X with odds-sucess rate given by the low-rank model tensor M

Distribution: $x_i \sim \operatorname{Bernouli}(\rho_i)$

Link function: $m_i = \frac{\rho_i}{1 - \rho_i}$

Loss function: $f(x, m) = \log(m + 1) - x\log(m + \epsilon)$

Domain: $m \in [0, \infty)$

GCPDecompositions.BetaDivergenceLoss — Type

BetaDivergenceLoss(β::Real, eps::Real)

BetaDivergence Loss for given β

Loss function: $f(x, m; β) = \frac{1}{\beta}m^{\beta} - \frac{1}{\beta - 1}xm^{\beta - 1} if \beta \in \mathbb{R} \{0, 1\}, m - x\log(m) if \beta = 1, \frac{x}{m} + \log(m) if \beta = 0$

Domain: $m \in [0, \infty)$

GCPDecompositions.CPD — Type

CPD

Tensor decomposition type for the canonical polyadic decompositions (CPD) of a tensor (i.e., a multi-dimensional array) A. This is the return type of gcp(_), the corresponding tensor decomposition function.

If M::CPD is the decomposition object, the weights λ and the factor matrices U = (U[1],...,U[N]) can be obtained via M.λ and M.U, such that A = Σ_j λ[j] U[1][:,j] ∘ ⋯ ∘ U[N][:,j].

GCPDecompositions.GammaLoss — Type

GammaLoss(eps::Real = 1e-10)

Loss corresponding to a statistical assumption of Gamma-distributed data X with scale given by the low-rank model tensor M.

Distribution: $x_i \sim \operatorname{Gamma}(k, \sigma_i)$

Link function: $m_i = k \sigma_i$

Loss function: $f(x,m) = \frac{x}{m + \epsilon} + \log(m + \epsilon)$

Domain: $m \in [0, \infty)$

GCPDecompositions.HuberLoss — Type

HuberLoss(Δ::Real)

Huber Loss for given Δ

Loss function: $f(x, m) = (x - m)^2 if \abs(x - m)\leq\Delta, 2\Delta\abs(x - m) - \Delta^2 otherwise$

Domain: $m \in \mathbb{R}$

GCPDecompositions.LeastSquaresLoss — Type

LeastSquaresLoss()

Loss corresponding to conventional CP decomposition. Corresponds to a statistical assumption of Gaussian data X with mean given by the low-rank model tensor M.

Distribution: $x_i \sim \mathcal{N}(\mu_i, \sigma)$

Link function: $m_i = \mu_i$

Loss function: $f(x,m) = (x-m)^2$

Domain: $m \in \mathbb{R}$

GCPDecompositions.NegativeBinomialOddsLoss — Type

NegativeBinomialOddsLoss(r::Integer, eps::Real = 1e-10)

Loss corresponding to the statistical assumption of Negative Binomial data X with log odds failure rate given by the low-rank model tensor M

Distribution: $x_i \sim \operatorname{NegativeBinomial}(r, \rho_i)$

Link function: $m = \frac{\rho}{1 - \rho}$

Loss function: $f(x, m) = (r + x) \log(1 + m) - x\log(m + \epsilon)$

Domain: $m \in [0, \infty)$

GCPDecompositions.NonnegativeLeastSquaresLoss — Type

NonnegativeLeastSquaresLoss()

Loss corresponding to nonnegative CP decomposition. Corresponds to a statistical assumption of Gaussian data X with nonnegative mean given by the low-rank model tensor M.

Distribution: $x_i \sim \mathcal{N}(\mu_i, \sigma)$

Link function: $m_i = \mu_i$

Loss function: $f(x,m) = (x-m)^2$

Domain: $m \in [0, \infty)$

GCPDecompositions.PoissonLogLoss — Type

PoissonLogLoss()

Loss corresponding to a statistical assumption of Poisson data X with log-rate given by the low-rank model tensor M.

Distribution: $x_i \sim \operatorname{Poisson}(\lambda_i)$

Link function: $m_i = \log \lambda_i$

Loss function: $f(x,m) = e^m - x m$

Domain: $m \in \mathbb{R}$

GCPDecompositions.PoissonLoss — Type

PoissonLoss(eps::Real = 1e-10)

Loss corresponding to a statistical assumption of Poisson data X with rate given by the low-rank model tensor M.

Distribution: $x_i \sim \operatorname{Poisson}(\lambda_i)$

Link function: $m_i = \lambda_i$

Loss function: $f(x,m) = m - x \log(m + \epsilon)$

Domain: $m \in [0, \infty)$

GCPDecompositions.RayleighLoss — Type

RayleighLoss(eps::Real = 1e-10)

Loss corresponding to the statistical assumption of Rayleigh data X with sacle given by the low-rank model tensor M

Distribution: $x_i \sim \operatorname{Rayleigh}(\theta_i)$

Link function: $m_i = \sqrt{\frac{\pi}{2}\theta_i}$

Loss function: $f(x, m) = 2\log(m + \epsilon) + \frac{\pi}{4}(\frac{x}{m + \epsilon})^2$

Domain: $m \in [0, \infty)$

GCPDecompositions.UserDefinedLoss — Type

UserDefinedLoss

Type for user-defined loss functions $f(x,m)$, where $x$ is the data entry and $m$ is the model entry.

Contains three fields:

func::Function : function that evaluates the loss function $f(x,m)$

deriv::Function : function that evaluates the partial derivative $\partial_m f(x,m)$ with respect to $m$

domain::Interval : Interval from IntervalSets.jl defining the domain for $m$

The constructor is UserDefinedLoss(func; deriv, domain). If not provided,

deriv is automatically computed from func using forward-mode automatic differentiation

domain gets a default value of Interval(-Inf, +Inf)

GCPDecompositions._checked_khatrirao_dims — Method

_checked_khatrirao_dims(A1, A2, ...)

Check that A1, A2, etc. have compatible dimensions for the Khatri-Rao product. If so, return a tuple of the number of rows and the shared number of columns. If not, throw an error.

GCPDecompositions._checked_mttkrp_dims — Method

_checked_mttkrp_dims(X, (U1, U2, ..., UN), n)

Check that X and U have compatible dimensions for the mode-n MTTKRP. If so, return a tuple of the number of rows and the shared number of columns for the Khatri-Rao product. If not, throw an error.

GCPDecompositions.create_mttkrp_buffer — Method

create_mttkrp_buffer(X, U, n)

Create buffer to hold intermediate calculations in mttkrp!.

Always use create_mttkrp_buffer to make a buffer for mttkrp!; the internal details of buffer may change in the future and should not be relied upon.

See also: mttkrp!

GCPDecompositions.gcp — Function

gcp(X::Array, r, loss = LeastSquaresLoss();
    constraints = default_constraints(loss),
    algorithm = default_algorithm(X, r, loss, constraints)) -> CPD

Compute an approximate rank-r CP decomposition of the tensor X with respect to the loss function loss and return a CPD object.

Keyword arguments:

constraints : a Tuple of constraints on the factor matrices U = (U[1],...,U[N]).

algorithm : algorithm to use

Conventional CP corresponds to the default LeastSquaresLoss() loss with the default of no constraints (i.e., constraints = ()).

If the LossFunctions.jl package is also loaded, loss can also be a loss function from that package. Check GCPDecompositions.LossFunctionsExt.SupportedLosses to see what losses are supported.

See also: CPD, AbstractLoss.

GCPDecompositions.khatrirao! — Method

khatrirao!(K, A1, A2, ...)

Compute the Khatri-Rao product (i.e., the column-wise Kronecker product) of the matrices A1, A2, etc. and store the result in K.

GCPDecompositions.khatrirao — Method

khatrirao(A1, A2, ...)

Compute the Khatri-Rao product (i.e., the column-wise Kronecker product) of the matrices A1, A2, etc.

GCPDecompositions.mttkrp! — Method

mttkrp!(G, X, (U1, U2, ..., UN), n, buffer=create_mttkrp_buffer(X, U, n))

Compute the Matricized Tensor Times Khatri-Rao Product (MTTKRP) of an N-way tensor X with the matrices U1, U2, ..., UN along mode n and store the result in G.

Optionally, provide a buffer for intermediate calculations. Always use create_mttkrp_buffer to make the buffer; the internal details of buffer may change in the future and should not be relied upon.

Algorithm is based on Section III-B of the paper:

Fast Alternating LS Algorithms for High Order CANDECOMP/PARAFAC Tensor Factorizations. Anh-Huy Phan, Petr Tichavský, Andrzej Cichocki. IEEE Transactions on Signal Processing, 2013. DOI: 10.1109/TSP.2013.2269903

See also: mttkrp, create_mttkrp_buffer

GCPDecompositions.mttkrp — Method

mttkrp(X, (U1, U2, ..., UN), n)

Compute the Matricized Tensor Times Khatri-Rao Product (MTTKRP) of an N-way tensor X with the matrices U1, U2, ..., UN along mode n.

See also: mttkrp!