GCPDecompositions

Documentation for GCPDecompositions.

GCPDecompositions.GCPDecompositions

GCPDecompositions.CPD

GCPDecompositions.gcp

GCPDecompositions.ncomponents

GCPDecompositions.GCPDecompositions — Module

Generalized CP Decomposition module. Provides approximate CP tensor decomposition with respect to general losses.

GCPDecompositions.CPD — Type

CPD

Tensor decomposition type for the canonical polyadic decompositions (CPD) of a tensor (i.e., a multi-dimensional array) A.

If F::CPD is the decomposition object, the weights λ and factor matrices U = (U[1],...,U[N]) can be obtained via F.λ and F.U, such that A = Σ_j λ[j] U[1][:,j] ∘ ⋯ ∘ U[N][:,j].

GCPDecompositions.gcp — Function

gcp(X::Array, r, func, grad, lower=-Inf) -> CPD

Compute an approximate rank-r CP decomposition of the tensor X with respect to a general loss and return a CPD object.

Inputs

X : multi-dimensional tensor/array to approximate/decompose

r : number of components for the CPD

func : loss function, default = (x, m) -> (m - x)^2

grad : loss function derivative, default = (x, m) -> 2 * (m - x)

lower : lower bound for factor matrix entries, default = -Inf

GCPDecompositions.ncomponents — Method

ncomponents(M::CPD) -> Integer

Return the number of components in M.

See also: ndims, size.